fixes for woudc according to the new check this adds the missing metadata fields
